requirements.txt
openai
chainlit
python_dotenv>=1.0.0, <2.0.0
pandas>=2.2.2, <3.0.0